adds first Navigation Route
