+++ .dart_tool/package_config.json
@@ -140,7 +140,7 @@
       "languageVersion": "2.7"
     }
   ],
-  "generated": "2020-10-07T19:34:14.721910Z",
+  "generated": "2020-10-08T18:11:18.918561Z",
   "generator": "pub",
   "generatorVersion": "2.9.2"
 }
