+++ .packages
@@ -1,4 +1,4 @@
-# Generated by pub on 2020-10-07 15:34:14.705258.
+# Generated by pub on 2020-10-08 14:11:18.903857.
 async:file:///Users/amyroberson/.pub-cache/hosted/pub.dartlang.org/async-2.4.2/lib/
 boolean_selector:file:///Users/amyroberson/.pub-cache/hosted/pub.dartlang.org/boolean_selector-2.0.0/lib/
 characters:file:///Users/amyroberson/.pub-cache/hosted/pub.dartlang.org/characters-1.0.0/lib/
